refactor(dossier): extract sections logic in a concern
