+++ app/models/dossier.rb
@@ -48,8 +48,9 @@
 class Dossier < ApplicationRecord
   self.ignored_columns = [:en_construction_conservation_extension]
   include DossierFilteringConcern
-  include DossierRebaseConcern
   include DossierPrefillableConcern
+  include DossierRebaseConcern
+  include DossierSectionsConcern
 
   enum state: {
     brouillon:       'brouillon',
@@ -1237,20 +1238,6 @@ class Dossier < ApplicationRecord
     termine_expired_to_delete.find_each(&:purge_discarded)
   end
 
-  def sections_for(champ)
-    @sections = Hash.new do |hash, parent|
-      case parent
-      when :public
-        hash[parent] = champs_public.filter(&:header_section?)
-      when :private
-        hash[parent] = champs_private.filter(&:header_section?)
-      else
-        hash[parent] = parent.champs.filter(&:header_section?)
-      end
-    end
-    @sections[champ.parent || (champ.public? ? :public : :private)]
-  end
-
   def clone
     dossier_attributes = [:autorisation_donnees, :user_id, :revision_id, :groupe_instructeur_id]
     relationships = [:individual, :etablissement]
@@ -1289,22 +1276,6 @@ class Dossier < ApplicationRecord
     return true if user_deleted?
 
     false
-  end
-
-  def auto_numbering_section_headers_for?(champ)
-    sections_for(champ)&.none?(&:libelle_with_section_index?)
-  end
-
-  def index_for_section_header(champ)
-    champs = champ.private? ? champs_private : champs_public
-
-    index = 1
-    champs.each do |c|
-      return index if c.stable_id == champ.stable_id
-      next unless c.visible?
-
-      index += 1 if c.type_de_champ.header_section?
-    end
   end
 
   private
